Update commendations_XCOMFILES.rul
Added proper entry for "Commended!" commendation , so agents can receive medal , bonus and most important (for me , perfectionist) ufopaedia would list that commendation as awarded. Criteria choosen is to incapacitate or kill at least 1 enemy. Why ? Since description mentioned "not bad for a rookie" , i choose something that would be criteria for rookie promotion (or close to it).